Remove files header. Split code into different files. Use ordered dictionary instead of array for events keeping.
