+++ LensInterface/LensBuilder.swift
@@ -1,10 +1,3 @@
-//
-//  LensInterfaceBuilder.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 25/7/22.
-//
-
 import Foundation
  
 class LensBuilder {
