+++ LensInterface/LensError.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct LensError: LocalizedError, Equatable {
+    
+    let errorDescription: String?
+    let failureReason: String?
+}
