+++ ReduxLens/Fake/FakeReducer.swift
@@ -1,10 +1,3 @@
-//
-//  FakeReducer.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 31/8/22.
-//
-
 import ComposableArchitecture
 import Foundation
 
@@ -12,7 +5,7 @@ let appReducer_fake = Reducer<AppState, AppAction, AppEnvironment> { state, acti
     switch action {
     case .setup:
         let fakeState = loadBigStateJson()
-        state.events = [(UUID(), ReduxEvent(fakeState, fakeState))]
+        state.events = [UUID(): ReduxEvent(fakeState, fakeState)]
         state.error = .navigationFailedLineNotFound
         return .none
     case .selectItem(id: let id):
@@ -33,7 +26,6 @@ extension ReduxEvent {
         _ stateAfter: [String: Any]
     ) {
         self.name = "Test action"
-        self.timestamp = Date().timeIntervalSince1970
         self.leadTime = Date().timeIntervalSince1970
         self.stateBefore = stateBefore
         self.stateAfter = stateAfter
