+++ ReduxLens/ReduxLensBuilder.swift
@@ -1,10 +1,3 @@
-//
-//  TAGABuilder.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 25/7/22.
-//
-
 import Foundation
 
 final class ReduxLensBuilder: LensBuilder {
