+++ ReduxLens/TCA/State/AppState.swift
@@ -1,0 +1,8 @@
+import Foundation
+import OrderedCollections
+
+struct AppState {
+    var events: OrderedDictionary<UUID, ReduxEvent> = [:]
+    var selectedItem: ReduxEvent? = nil
+    var error: ReduxError?
+}
