+++ ReduxLens/TCA/State/ReduxError.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum ReduxError: Error {
+    case navigationFailedFileNotFound
+    case navigationFailedLineNotFound
+    case eventNotDeserialiazable(DecodingError)
+}
