+++ Spyglass/LensLoader.swift
@@ -1,10 +1,3 @@
-//
-//  LensLoader.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 25/7/22.
-//
-
 import Foundation
 
 typealias InitFunction = @convention(c) () -> UnsafeMutableRawPointer
