+++ Spyglass/SpyglassApp.swift
@@ -1,10 +1,3 @@
-//
-//  SpyglassApp.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 25/7/22.
-//
-
 import SwiftUI
 
 @main
