+++ Spyglass/View/CatalogView.swift
@@ -1,10 +1,3 @@
-//
-//  CatalogView.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 1/9/22.
-//
-
 import SwiftUI
 
 struct CatalogView: View {
