+++ Spyglass/View/ContentView.swift
@@ -1,10 +1,3 @@
-//
-//  ContentView.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 25/7/22.
-//
-
 import SwiftUI
 
 struct ContentView: View {
