+++ Spyglass/View/SearchField.swift
@@ -1,10 +1,3 @@
-//
-//  TextDebouncer.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 1/9/22.
-//
-
 import SwiftUI
 
 struct SearchField: View {
