+++ Spyglass/View/TreeView.swift
@@ -1,10 +1,3 @@
-//
-//  TreeView.swift
-//  Spyglass
-//
-//  Created by Aleksey Yakimenko on 31/8/22.
-//
-
 import AppKit
 import OutlineView
 import SwiftUI
